+++ packages/collaborative/src/collaborative/schema/note.schema.ts
