+++ packages/collaborative/src/collaborative/schema/space.schema.ts
