--- packages/collaborative/src/note/note.module.ts
@@ -1,16 +1,0 @@
-import { Module } from '@nestjs/common';
-import { MongooseModule } from '@nestjs/mongoose';
-
-import { NoteDocument, NoteSchema } from './note.schema';
-import { NoteService } from './note.service';
-
-@Module({
-  imports: [
-    MongooseModule.forFeature([
-      { name: NoteDocument.name, schema: NoteSchema },
-    ]),
-  ],
-  providers: [NoteService],
-  exports: [NoteService],
-})
-export class NoteModule {}
