--- packages/collaborative/src/space/space.module.ts
@@ -1,16 +1,0 @@
-import { Module } from '@nestjs/common';
-import { MongooseModule } from '@nestjs/mongoose';
-
-import { SpaceDocument, SpaceSchema } from './space.schema';
-import { SpaceService } from './space.service';
-@Module({
-  imports: [
-    MongooseModule.forFeature([
-      { name: SpaceDocument.name, schema: SpaceSchema },
-    ]),
-  ],
-  providers: [SpaceService],
-
-  exports: [SpaceService],
-})
-export class SpaceModule {}
